+++ backend/src/sale.module.ts
@@ -2,7 +2,8 @@ import { Module } from '@nestjs/common';
 import { SaleController } from './presentation/sale/controllers/sale.controller';
 import { SaleService } from './application/sale/services/sale.service';
 import { SalePrismaRepository } from './infrastructure/sale/repositories/sale.prisma.repository';
-import { SALE_REPOSITORY } from './domain/sale/interfaces/sale.interface';
+import { SaleHistoryPrismaRepository } from './infrastructure/sale/repositories/sale-history.prisma.repository';
+import { SALE_REPOSITORY, SALE_HISTORY_REPOSITORY } from './domain/sale/interfaces/sale.interface';
 import { PRODUCT_REPOSITORY } from './domain/product/interfaces/product.interface';
 import { ProductPrismaRepository } from './infrastructure/product/repositories/product.prisma.repository';
 import { PACKAGE_REPOSITORY, SALE_PACKAGE_REPOSITORY } from './domain/package/interfaces/package.interface';
@@ -18,6 +19,7 @@ import { TerceroPrismaRepository } from './infrastructure/tercero/repositories/t
   providers: [
     SaleService,
     { provide: SALE_REPOSITORY, useClass: SalePrismaRepository },
+    { provide: SALE_HISTORY_REPOSITORY, useClass: SaleHistoryPrismaRepository },
     { provide: PRODUCT_REPOSITORY, useClass: ProductPrismaRepository },
     { provide: PACKAGE_REPOSITORY, useClass: PackagePrismaRepository },
     { provide: SALE_PACKAGE_REPOSITORY, useClass: SalePackagePrismaRepository },